Add graphical user interface with new styles
Added a new graphical user interface with specific styles defined in a new css file. These styles include the definition of color schemes, fonts, hover effects, media queries, and others. Also, the creation of SVG files for logos and the implementation of certain dependencies in Cargo.lock file for the new interface under the 'gui' directory.