Refactor code and introduce new CropProcessor
Implemented a geometric CropProcessor in the processing area of the project. The processors/geometry folder added, contains the implementation of the CropProcessor. Several other files have been affected as part of a general code refactor and organization. In addition, updates were made to the FastImage usage throughout the file changes.