Refactor end-to-end testing structure for CLI
Move e2e tests to the `cli` crate and restructure unit tests. The end-to-end testing code has been moved directly under the relevant `cli` crate so that dependencies only occur in the relevant places removing the need for an extra `cli-e2e` package. The handling of test files has been improved, with the introduction of a `TestFiles` struct. Changes also include additional dependencies update in the Cargo.toml