Refactor code to replace ArgType with PicturifyArg
This refactor aims to improve the maintainability of the code by replacing ArgType with the more flexible PicturifyArg. This allows the definition of parameters to be more readable and easier to handle. Additional improvements include the extraction of default argument values into constants, enhancement of command structures, and minor changes to FastImage methods.